Refactor based on static analysis using tools like CheckStyle, SonarQube, and SpotBugs.

+++ src/main/java/org/asdfjkl/jfxchess/lib/OptimizedRandomAccessFile.java
@@ -317,7 +317,7 @@ public class OptimizedRandomAccessFile {
                     nextChar++;
                 }
             }
-            long d = nChars - nextChar;
+            long d = (long) nChars - nextChar;
             if (r <= d) {
                 nextChar += r;
                 r = 0;
